Refactor: vote-at-home separate components
