+++ src/components/Home/VoteAtHome/VoteCard/CardNull/CardNull.module.scss
@@ -1,0 +1,67 @@
+@use "@/sass";
+
+.voteBox {
+  height: 175px;
+
+  background: linear-gradient(
+    119deg,
+    #62aaff 35.27%,
+    rgba(98, 170, 255, 0.3) 142.38%
+  );
+
+  border-radius: 16px;
+
+  .contents {
+    height: 100%;
+
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+
+    padding: 33px 24px;
+
+    .left {
+      height: 100%;
+
+      display: flex;
+      flex-direction: column;
+      justify-content: space-between;
+
+      .text {
+        @include sass.typography(titleSmall);
+      }
+    }
+    .right {
+      width: 98px;
+      height: 98px;
+      padding: 5px 13px;
+    }
+  }
+  .contents {
+    height: 100%;
+
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+
+    padding: 33px 24px;
+
+    .left {
+      height: 100%;
+
+      display: flex;
+      flex-direction: column;
+      justify-content: space-between;
+
+      .text {
+        @include sass.typography(titleSmall);
+        color: sass.$neutral0;
+      }
+    }
+    .right {
+      width: 98px;
+      height: 98px;
+      padding: 5px 13px;
+    }
+  }
+}
